requirements.txt
mysql-connector-python==8.0.27
python-dotenv==0.19.2
Flask==1.0.2
uWSGI==2.0.17.1
Werkzeug==0.14.1
